requirements.txt
streamlit==1.31.0
chromadb==0.4.22
openai>=1.30.0
python-dotenv==1.0.1
sentence-transformers==2.3.1
torch==2.1.0